- Note is now an Entity - PSQL Database connected to App - Table created via QueryConsole - noteRepository changed to JpaRepository (name is still NoteRepository) - adapted NoteService to new JpaRepository - implemented DataJpaTests for new NoteRepository (and it works, yay) - the other Tests need an overhaul and don't work yet

+++ Server (back-end)/to-do-list/src/main/java/com/example/todolist/controller/NoteController.java
@@ -1,7 +1,7 @@
 package com.example.todolist.controller;
 
 import com.example.todolist.logic.NoteService;
-import com.example.todolist.persistence.data.Note;
+import com.example.todolist.persistence.model.Note;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
@@ -32,7 +32,7 @@ public class NoteController {
     }
 
     @DeleteMapping("/{id}")
-    void deleteNote(@PathVariable int id) {
+    void deleteNote(@PathVariable Long id) {
         noteService.deleteNote(id);
     }
 }
